==> 08-FHBench.Rmd <==
```{r setup, include=FALSE, message=FALSE, error=FALSE, warning=FALSE}
knitr::opts_chunk$set(warning = FALSE, 
                      message = FALSE,
                      cache = TRUE)

library(printr)
library(kableExtra)
library(tidyverse)
library(magrittr)
theme_set(theme_bw())
tba <- function(dat, cap = NA){
  kable(dat,
      format = "html", digits =  4,
      caption = cap) %>% 
     kable_styling(bootstrap_options = "striped", full_width = F)%>%
         kable_classic(full_width = F, html_font = "Arial Narrow")
}
```

# Benchmark regional para las estimaciones del modelo de área

## Lectura de librerías

```{r}
library(survey)
library(srvyr)
library(TeachingSampling)
library(stringr)
library(magrittr)
library(sae)
library(ggplot2)
library(emdi)
library(patchwork)
library(DBI)
library(odbc)
library(flexmix)
library(modeltools)
library(sp)
library(sf)
library(rgdal)
library(tmap)
library(dplyr)
id_dominio <- "id_dominio"
```

## Lectura de bases de datos 

El código está leyendo varios archivos necesarios para el análisis.

  - `poligonos_dominios`: Es un archivo shapefile que contiene la información geográfica (polígonos) de los dominios de la encuesta. Se carga usando la función `read_sf()` del paquete `sf`.
  -   `base_completa`: Es un archivo RDS que contiene la información de la encuesta. Incluye información de los dominios, los estratos, los conglomerados y los pesos.
  -   `estimacionesPre`: Es un archivo RDS que contiene las estimaciones directas de los dominios.
  -   `base_FH`: Es una base que se crea al juntar la `base_completa` con las `estimacionesPre` usando la función `left_join()` del paquete `dplyr`.
  -   `encuesta`: Es un archivo RDS que contiene la información de la encuesta. 
  -   `fh_arcsin`: Es un archivo RDS que contiene las estimaciones del modelo ajustado de Fay-Herriot.
  -   `personas_dominio`: Es un archivo RDS que contiene número de personas en cada dominio.
  -   `personas_dominio_agregado`: Es un archivo RDS que contiene el número total de personas agregada por región.

```{r}
## Shapefile
poligonos_dominios <- read_sf( "../shapefiles2010/DOM.shp") 

## Base de estimaciones 
base_completa <- readRDS('../Data/base_completa.Rds')
estimacionesPre <- readRDS('../Data/estimaciones.Rds')
base_FH <- left_join(base_completa, estimacionesPre,   by = id_dominio)

## Encuesta
encuesta <- readRDS("../Data/encuestaDOM.Rds")

## Estimaciones del modelo ajustado
fh_arcsin <- readRDS("../Data/fh_arcsin.Rds")

## Agregados por dominios y región 
personas_dominio <- readRDS('../Data/agregado_persona_dominio.rds') 
personas_dominio_agregado <- readRDS('../Data/agregado_persona_region.rds') 
```

## Estimación directa por región 

El código carga y prepara los datos de la encuesta y luego calcula las estimaciones directas en cada región del país.

Primero, para el conjunto de datos de la encuesta se utiliza la función `mutate()` de `dplyr` para agregar tres columnas: `upm`, `estrato` y `factor_anual`. La columna `upm` es la identificación única de la unidad primaria de muestreo, `estrato` es la identificación del estrato en el que se encuentra la unidad primaria de muestreo y `factor_anual` es el factor de expansión anual, calculado dividiendo el factor de expansión trimestral entre 4.

A continuación, se utiliza la función `as_survey_design()` de la librería `survey` para crear un objeto de diseño de la encuesta a partir de los datos. La función utiliza las columnas `estrato`, `upm` y `factor_anual` para especificar las estrategias de muestreo, el peso de muestreo y la estructura de conglomerados.

Por último, se utiliza la función `group_by()` de `dplyr` para agrupar los datos por `id_region`, luego se utiliza la función `summarise()` de `dplyr` junto con `survey_ratio()` de `survey` para calcular la estimación directa. El resultado se asigna a un objeto llamado `directoDepto`, que contiene las columnas `id_region` y `theta_region`, que corresponde a la estimación directa en cada región.

```{r}
encuesta <-
  encuesta %>% 
  mutate(
    upm = str_pad(string = upm,width = 9,pad = "0"),
    estrato = str_pad(string = estrato,width = 5,pad = "0"),
    factor_anual = factor_expansion / 4
  ) %>% data.frame()

disenoDOM <- encuesta %>%
  as_survey_design(
    strata = estrato,
    ids = upm,
    weights = factor_anual,
    nest=T
  )

directoDepto <- disenoDOM %>%
  group_by(id_region) %>% 
  filter(ocupado == 1 & pet == 1) %>%
  summarise(Rd = survey_ratio(
    numerator = orden_sector == 2 ,
    denominator = 1,
    vartype = c("se", "ci", "var", "cv"),
    deff = T
  )) %>% 
  transmute(id_region, theta_region = Rd)

```

## Consolidación de la base de datos 

Este código realiza los siguientes pasos:

  -   Realiza un `left_join()` entre la tabla `directoDepto` y la tabla `personas_dominio` por la columna `id_region`. 
  -   Realiza un `left_join()` entre la tabla resultante del paso anterior y la tabla `personas_dominio_agregado` por la columna `id_region`. El resultado de esta unión es una tabla que agrega la información por dominio a nivel de región.
  
  -   Realiza un `left_join()` entre la tabla resultante del paso anterior y la tabla `estimacionesPre` por la columna `id_dominio`. El resultado de esta unión es una tabla que agrega la información de ambas tablas por dominio.
  -   Transmuta la columna `FayHerriot` de la tabla resultante del paso anterior y cambia su nombre a `FayHarriot`.
El resultado final es una tabla llamada `R_mpio` que contiene información agregada por región y dominio, así como las estimaciones por dominio utilizando el método de Fay Herriot.

```{r}
R_mpio <- directoDepto %>% 
  left_join(personas_dominio, by = 'id_region') %>%
  left_join(personas_dominio_agregado, by = "id_region") %>%
  left_join(estimacionesPre %>% 
              transmute(id_dominio, FayHerriot = FH),
            by = id_dominio) %>% 
  data.frame()

```


## Pesos Benchmark
Para obtener el factor multiplicador de de Benchmarking se ejecuta el código siguiente. 
El código está creando un objeto llamado `R_mpio2`, el cual se está construyendo a partir del objeto `R_mpio`. Primero, se agrupa por `id_region` usando `group_by()` y se calculan dos medidas de razón de la encuesta utilizando `summarise()`. Los cálculos de las medidas de razón no están claros ya que dependen de variables no presentes en el código. Luego, se realiza un `left_join()` con el objeto `directoDepto` para añadir la variable `theta_region`.

Posteriormente, se construye un objeto llamado pesos a partir del objeto `R_mpio` utilizando `ungroup()` para eliminar la agrupación anteriormente realizada. Luego, se añade la columna `W_i` que representa los pesos muestrales utilizados para el análisis.

La función `tba()` no está disponible en R base o en los paquetes `dplyr`, por lo que no es posible determinar lo que hace. Probablemente es una función personalizada por el autor del código.

```{r}
R_mpio2 <- R_mpio %>% 
  group_by(id_region) %>% 
  summarise(
    R_region_RB = unique(theta_region) / sum((pp_dominio  / pp_region ) * FayHerriot),
    R_region_DB = unique(theta_region) - sum((pp_dominio  / pp_region ) * FayHerriot)
  ) %>%
  left_join(directoDepto, by = "id_region")

pesos <- R_mpio %>% ungroup() %>% 
  mutate(W_i = pp_dominio / pp_region) %>% 
  dplyr::select(id_dominio, W_i)

head(pesos, 15) %>% tba()

```

## Estimación FH Benchmark 

El siguiente código realiza una serie de operaciones sobre dos `data.frames`: `estimacionesPre` y `R_mpio` para obtener las estimaciones ajustadas mediante el benchmarking.

En primer lugar, se realiza un `left_join` entre `estimacionesPre` y `R_mpio` utilizando la columna `id_dominio`. 

Luego, se realiza un nuevo `left_join` entre este `data.frame` resultante y `R_mpio2`, utilizando la columna `id_region`. Este join agrega las columnas `R_region_RB` y `R_region_DB` de `R_mpio2` al `data.frame()`.

A continuación, se calcula una nueva columna llamada `FH_RBench` que resulta de multiplicar la columna FH por la columna `R_region_RB`.

Por último, se realiza un join entre este `data.frame` resultante y pesos, utilizando la columna `id_dominio`. Este join agrega la columna `W_i` de `pesos` al `data.frame`. El resultado final se guarda en la variable `estimacionesBench`.

```{r}
estimacionesBench <- estimacionesPre %>% 
  left_join(R_mpio %>% 
              dplyr::select(id_region, id_dominio), 
            by = id_dominio) %>%
  left_join(R_mpio2, by = c("id_region")) %>% ungroup() %>% 
  mutate(FH_RBench = R_region_RB * FH) %>%
  left_join(pesos, by = id_dominio)
```

### Validación: Estimación FH con Benchmark 
Al realizar el proceso de validación notamos que las estimaciones con directas de las regiones y las estimaciones de FH con Benchmarking  son iguales. 
```{r}
estimacionesBench %>% group_by(id_region) %>%
  summarise(theta_reg_RB = sum(W_i * FH_RBench)) %>%
  left_join(directoDepto, by = "id_region") %>%
  tba()
```

### Comparación gráfica entre estimación FH y FH con Benchmark por dominio 

El código presentado realiza una comparación de las estimaciones Fay-Herriot originales con las estimaciones Fay-Herriot con benchmark.


```{r}
temp_Bench <- estimacionesBench %>%
  transmute(
    id_dominio,
    Directo = Direct * 100,
    FayHerriot = FH * 100,
    FH_RBench = FH_RBench * 100,
    Gamma)

theme_set(theme_bw())
ggplot(temp_Bench, aes(FH_RBench, FayHerriot)) + 
  geom_point() + geom_abline(intercept = 0, slope = 1) + 
  labs(y = "Estimación Fay-Herriot",
       x = "Estimación Fay-Herriot con Benchmark")

```

## Tabla final de estimaciones 

### Función para el calculo de intervalos de confianza 

La función `ICL()` sirve para calcular intervalos de confianza para la proporción de una población, dada una muestra y un nivel de confianza. Los argumentos de la función son:

  -   `p`: la proporción muestral
  -   `mse`: el error cuadrático medio
  -   `alpha`: el nivel de confianza deseado (por defecto es 0.05)
  -   `student`: un indicador booleano que indica si se debe usar una distribución t de Student en lugar de la distribución normal para el cálculo del intervalo de confianza (por defecto es FALSE)
  -   `nu`: el número de grados de libertad si se usa una distribución t de Student (por defecto es NULL y se calcula automáticamente)

```{r}
ICL <- function(p, mse, alpha = 0.05, student = FALSE, nu = NULL) {
  if (student == TRUE) {
    q <- qt(1 - alpha/2, nu)
  } else {
    q <- qnorm(1 - alpha/2)
  }
  CL <- log(p/(1 - p)) - (q * sqrt(mse))/(p * (1 - p))
  CU <- log(p/(1 - p)) + (q * sqrt(mse))/(p * (1 - p))
  IC_1 <- exp(CL)/(1 + exp(CL))
  IC_2 <- exp(CU)/(1 + exp(CU))
  
  return(data.frame(L.I = IC_1, L.S = IC_2))
}
```

La función `TablaFinal` toma como entrada la tabla `estimacionesBench` y realiza una serie de transformaciones y cálculos para generar una tabla final con las estimaciones y los intervalos de confianza correspondientes.

Primero, agrega dos columnas nuevas: `sintetico` y `sintetico_back`. La columna `sintetico` se obtiene a partir de una matriz generada a partir de la columna `base_FH` de la tabla `estimacionesBench`, que contiene los valores de las variables auxiliares utilizadas en el modelo de Fay-Herriot. Esta matriz se multiplica por los coeficientes del modelo ajustado, que se asume que se ha generado previamente. La columna `sintetico_back` es el resultado de aplicar una transformación inversa al valor de `sintetico`.

Luego, se seleccionan y renombran las columnas que se incluirán en la tabla final. Estas columnas incluyen la estimación directa (`Directo`), el error estándar de la estimación directa (`ee_directo`), el coeficiente de variación de la estimación directa (`CV_directo`), la estimación de Fay-Herriot (`FayHerriot`), la raíz del error cuadrático medio de la estimación de Fay-Herriot (`rmse_FH`), el coeficiente de variación de la estimación de Fay-Herriot (`rrmse_FH`), la estimación de  gamma (`Gamma`), las estimaciones sintéticas (`sintetico` y `sintetico_back`), la estimación de Fay-Herriot con el ajuste de benchmark (`FH_RBench`), y los límites inferior y superior de los intervalos de confianza calculados utilizando el método de la distribución normal (`LI_normal` y `LS_normal`) y el método logit (`LI_logit` y `LS_logit`). Finalmente, se calculan los límites inferior y superior de los intervalos de confianza finales (`LI_final` y `LS_final`), teniendo en cuenta los límites de 0 y 1.

```{r}
TablaFinal <- estimacionesBench %>%
  mutate(
    sintetico = as.matrix(base_FH %>% data.frame() %>% 
                            dplyr::select(rownames(
                              fh_arcsin$model$coefficients
                            ))) %*%
      fh_arcsin$model$coefficients[, 1],
    sintetico_back = sin(sintetico) ^ 2
  ) %>%
  transmute(
    id_dominio,
    n_muestral = n,
    Directo = Direct,
    ee_directo = sqrt(Direct_MSE),
    CV_directo = Direct_CV,
    FayHerriot = FH,
    rmse_FH = sqrt(FH_MSE),
    rrmse_FH = rmse_FH / FayHerriot,
    Gamma,
    sintetico,
    sintetico_back,
    FH_RBench,
    LI_normal = FH_RBench - 1.96 * sqrt(FH_MSE),
    LS_normal = FH_RBench + 1.96 * sqrt(FH_MSE),
    LI_logit = ICL(FH_RBench, FH_MSE)[, 1],
    LS_logit = ICL(FH_RBench, FH_MSE)[, 2],
    LI_final = ifelse(LI_normal < 0, LI_logit, LI_normal),
    LS_final = ifelse(LS_normal > 1, LS_logit, LS_normal)
  )
 tba(head(TablaFinal,20))

```

Comparando los limites superior e inferior de los IC

```{r}

a1 <- ggplot(TablaFinal, aes(x = LI_normal, y = LI_logit)) +
  geom_point() +
  geom_abline(aes(intercept = 0, slope = 1), col = 2)  +
  labs(y = "LI_logit", x = "LI")

a2 <- ggplot(TablaFinal, aes(x = LS_normal, y = LS_logit)) +
  geom_point() +
  geom_abline(aes(intercept = 0, slope = 1), col = 2)  +
  labs(y = "LS_logit", x = "LS")

a1 | a2

```

Guardar resultados 

```{r, eval=FALSE}
saveRDS(TablaFinal, '../Data/TablaFinal.Rds')
saveRDS(estimacionesBench, '../Data/estimacionesBench.Rds')
```

## Creando mapa de resultados 

La función `tm_shape()` de la librería `tmap` se utiliza para especificar los datos espaciales (mapa) que se van a `graficar`. En este caso, se le pasa como argumento la tabla `poligonos_dominios`.

La función `tm_fill()` se utiliza para definir el estilo de relleno de los polígonos en el mapa. En este caso, se utiliza para graficar la variable `fh_porc`. Se utiliza el estilo "quantile" para que se dividan los valores en cuantiles y se apliquen diferentes tonalidades de color para cada cuantil.

La función `tm_borders()` se utiliza para especificar las propiedades de las líneas que delinean los polígonos.

La función `tm_layout()` se utiliza para definir los elementos del diseño del mapa, como la posición y estilo de la leyenda. En este caso, se define que la leyenda se ubique en la parte inferior central del mapa, que el fondo de la leyenda sea blanco con transparencia del 10%, y se especifica el tamaño del texto de la leyenda.

```{r}
poligonos_dominios <- poligonos_dominios %>%
  left_join( temp_Bench, by = id_dominio ) %>%
  mutate( fh_porc = FH_RBench )

tmap_options(check.and.fix = TRUE)
mapa <- tm_shape( poligonos_dominios ) +
  tm_fill( "fh_porc", style = "quantile", title="Tasa de informalidad" ) +
  tm_borders( col = "black", lwd=1, lty = "solid") +
  tm_layout( #"Wealth (or so)",
    legend.title.size = 1,
    legend.text.size = 0.6,
    legend.position = c( "center","bottom" ),
    legend.bg.color = "white",
    legend.stack = "horizontal",
    #legend.digits = 5,
    legend.bg.alpha = 0.1) 
mapa

```
